agentforge/config.py
"""Central configuration — loads env vars and defines constants."""

from __future__ import annotations

import os
from pathlib import Path

from dotenv import load_dotenv

load_dotenv()

# ── Paths ──────────────────────────────────────────────────────────────
PROJECT_ROOT = Path(__file__).resolve().parent.parent
DATA_DIR = PROJECT_ROOT / "data"
DATA_DIR.mkdir(parents=True, exist_ok=True)

# ── LLM (OpenRouter) ───────────────────────────────────────────────────
OPENROUTER_API_KEY: str = os.getenv("OPENROUTER_API_KEY", "")
OPENROUTER_BASE_URL: str = os.getenv("OPENROUTER_BASE_URL", "https://openrouter.ai/api/v1")
MODEL_NAME: str = os.getenv("MODEL_NAME", "google/gemini-2.0-flash-001")
TEMPERATURE: float = float(os.getenv("TEMPERATURE", "0.1"))

# ── Database ───────────────────────────────────────────────────────────
DB_PATH: str = os.getenv("DB_PATH", str(DATA_DIR / "agentforge.db"))

# ── ChromaDB ───────────────────────────────────────────────────────────
CHROMA_PERSIST_DIR: str = os.getenv("CHROMA_PERSIST_DIR", str(DATA_DIR / "chroma_data"))
CHROMA_COLLECTION: str = os.getenv("CHROMA_COLLECTION", "agentforge_codebase")

# ── Security ───────────────────────────────────────────────────────────
# Comma-separated API keys; empty = auth disabled (dev mode)
_raw_keys = os.getenv("API_KEYS", "")
API_KEYS: set[str] = {k.strip() for k in _raw_keys.split(",") if k.strip()}

RATE_LIMIT_RPM: int = int(os.getenv("RATE_LIMIT_RPM", "60"))  # 0 = disabled
MAX_CODE_SIZE: int = int(os.getenv("MAX_CODE_SIZE", str(512 * 1024)))  # 512KB

# Comma-separated CORS origins; * = allow all (dev mode)
CORS_ORIGINS: list[str] = os.getenv("CORS_ORIGINS", "*").split(",")

# ── Logging ────────────────────────────────────────────────────────────
LOG_LEVEL: str = os.getenv("LOG_LEVEL", "INFO")

# ── GitHub Integration ─────────────────────────────────────────────────
GITHUB_TOKEN: str = os.getenv("GITHUB_TOKEN", "")
GITHUB_WEBHOOK_SECRET: str = os.getenv("GITHUB_WEBHOOK_SECRET", "")

# ── Agent System Prompts ───────────────────────────────────────────────

ROUTER_SYSTEM_PROMPT = """\
You are a code classification expert. Given a code snippet, you must analyze it and return a structured routing decision.

Determine:
1. **language** — The programming language (python, javascript, typescript, java, go, rust, etc.)
2. **domain** — The code domain (web, data, infrastructure, ml, general)
3. **complexity** — Code complexity level (low, medium, high)
4. **risk_level** — Risk level for potential issues (low, medium, high)
5. **agents_to_activate** — Which review agents should analyze this code. Choose from: ["security", "performance", "architecture", "correctness"]. Always include "correctness". Include "security" for web/API/auth code. Include "performance" for data-heavy or algorithmic code. Include "architecture" for complex multi-class/module code.

Be precise and return ONLY the structured JSON output."""

SECURITY_SYSTEM_PROMPT = """\
You are an elite security code reviewer. Analyze the provided code for security vulnerabilities.

Focus on:
- **Injection attacks**: SQL injection, command injection, XSS, SSTI
- **Authentication & Authorization flaws**: Missing auth checks, weak token handling
- **Secret exposure**: Hardcoded API keys, passwords, tokens in code
- **Insecure deserialization**: Pickle, eval, exec usage
- **Dependency vulnerabilities**: Known insecure patterns
- **Input validation**: Missing or insufficient validation
- **Cryptographic issues**: Weak hashing, insecure random generation

For each finding, provide:
- severity: "critical", "warning", or "info"
- category: the security sub-category
- description: clear explanation of the vulnerability
- line_start / line_end: approximate line range (0 if unknown)
- suggested_fix: concrete code fix or mitigation

Be thorough but avoid false positives. Only report genuine security concerns."""

PERFORMANCE_SYSTEM_PROMPT = """\
You are an expert performance code reviewer. Analyze the provided code for performance issues.

Focus on:
- **Time complexity**: O(n²) or worse algorithms where O(n log n) or better exists
- **Space complexity**: Unnecessary memory allocations, large object copies
- **N+1 query patterns**: Database queries inside loops
- **Missing caching**: Repeated expensive computations
- **Async optimization**: Blocking calls that should be async
- **Resource leaks**: Unclosed file handles, connections, cursors
- **Unnecessary work**: Redundant computations, unused results

For each finding, provide:
- severity: "critical", "warning", or "info"
- category: the performance sub-category
- description: clear explanation of the issue with Big-O analysis where applicable
- line_start / line_end: approximate line range (0 if unknown)
- suggested_fix: optimized code or approach

Focus on impactful issues, not micro-optimizations."""

ARCHITECTURE_SYSTEM_PROMPT = """\
You are a senior software architect reviewing code for design quality.

Focus on:
- **SOLID principles**: Single responsibility, open/closed, Liskov substitution, interface segregation, dependency inversion
- **Design patterns**: Appropriate (or missing) use of patterns
- **Coupling & Cohesion**: Tight coupling between modules, low cohesion within classes
- **API design**: Unclear interfaces, inconsistent naming, poor abstractions
- **Code organization**: God classes, feature envy, inappropriate intimacy
- **Error handling design**: Inconsistent error handling strategies
- **Naming conventions**: Unclear or misleading names

For each finding, provide:
- severity: "critical", "warning", or "info"
- category: the architecture sub-category
- description: clear explanation referencing specific design principles
- line_start / line_end: approximate line range (0 if unknown)
- suggested_fix: refactored approach or design improvement

Focus on structural issues, not style nitpicks."""

CORRECTNESS_SYSTEM_PROMPT = """\
You are a meticulous code reviewer focused on logical correctness.

Focus on:
- **Logic errors**: Incorrect conditions, wrong operators, flawed algorithms
- **Edge cases**: Off-by-one errors, empty inputs, boundary conditions
- **Null/None handling**: Missing null checks, potential NoneType errors
- **Type mismatches**: Incompatible types, implicit conversions
- **Error handling**: Missing try/except, swallowed exceptions, wrong exception types
- **Concurrency bugs**: Race conditions, deadlocks, thread-safety issues
- **State management**: Uninitialized variables, stale state, mutation bugs

For each finding, provide:
- severity: "critical", "warning", or "info"
- category: the correctness sub-category
- description: clear explanation of the bug with specific scenario
- line_start / line_end: approximate line range (0 if unknown)
- suggested_fix: corrected code

Be precise — only report real bugs, not style preferences."""

SYNTHESIZER_SYSTEM_PROMPT = """\
You are a senior engineering lead synthesizing multiple code review reports into a single, actionable review.

You will receive reports from specialized agents (security, performance, architecture, correctness). Your job is to:

1. **Deduplicate**: Merge findings that refer to the same issue from different angles.
2. **Resolve conflicts**: When agents disagree (e.g., security recommends one approach, performance another), use your judgment to recommend the best trade-off and explain why.
3. **Prioritize**: Order findings by impact — critical issues first, then warnings, then informational.
4. **Score**: Assign an overall code quality score from 0 to 100 based on:
   - 90-100: Production-ready, minor suggestions only
   - 70-89: Good quality, some improvements recommended
   - 50-69: Needs work, significant issues found
   - 0-49: Major issues, requires substantial revision
5. **Summarize**: Write a clear executive summary of the code's strengths and weaknesses.

Return a structured review with the deduplicated findings, resolved conflicts, overall score, and summary."""
