pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "Multi-Agent Code Review & Debugging Platform"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.0",
]

[project.scripts]
agentforge = "agentforge.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ── Ruff ───────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "S101",   # assert allowed in tests
    "S603",   # subprocess calls OK in controlled contexts
    "T201",   # print OK in CLI
    "E501",   # line length handled by formatter
    "UP042",  # str+Enum needed for Pydantic JSON compat
    "SIM117", # nested with is clearer in test mocking
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "S106"]
"samples/*" = ["S", "B", "N", "SIM", "T20"]

# ── Pytest ─────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
